# src/index.js
const express = require('express');
const { v4: uuidv4 } = require('uuid');
const { serve, setup } = require('swagger-ui-express');
const swaggerFile = require('./swagger.json');

const PORT = 3333;
const app = express();
const customers = [];

app.use(express.json());
app.use('/swagger', serve, setup(swaggerFile));

function verifyAccount(request, response, next) {
  const { cpf } = request.headers;

  const hasCustomer = customers.find(customer => customer.cpf === cpf);

  if (!hasCustomer) {
    return response.status(400).json({ error: 'Customer not found' });
  }

  request.customer = hasCustomer;

  return next();
}

function getBalance(statement) {
  const balance = statement.reduce((acc, operation) => {
    if (operation.type === 'credit') {
      return acc + operation.amount;
    }
    return acc - operation.amount;
  }, 0);
  return balance;
}

app.post('/account', (request, response) => {
  const { cpf, name } = request.body;

  const hasCustomer = customers.some(customer => customer.cpf === cpf);

  if (hasCustomer) {
    return response.status(400).json({ error: 'Customer already exists!' });
  }

  customers.push({
    id: uuidv4(),
    cpf,
    name,
    statement: [],
  });

  return response.status(201).send();
});

app.put('/account', verifyAccount, (request, response) => {
  const { name } = request.body;
  const { customer } = request;

  customer.name = name;
  return response.status(200).send();
});

app.get('/account', verifyAccount, (request, response) => {
  const { customer } = request;

  return response.json(customer);
});

app.delete('/account', verifyAccount, (request, response) => {
  const { customer } = request;

  customers.splice(customer, 1);

  return response.status(200).json(customers);
});

app.post('/deposit', verifyAccount, (request, response) => {
  const { description, amount } = request.body;
  const { customer } = request;

  const statementOperation = {
    description,
    amount,
    created_at: new Date(),
    type: 'credit',
  };

  customer.statement.push(statementOperation);

  return response.status(201).send();
});

app.post('/withdraw', verifyAccount, (request, response) => {
  const { amount } = request.body;
  const { customer } = request;

  const balance = getBalance(customer.statement);

  if (balance < amount) {
    return response.status(400).json({ error: 'Insufficient funds !' });
  }

  const statementOperation = {
    amount,
    created_at: new Date(),
    type: 'debit',
  };

  customer.statement.push(statementOperation);

  return response.status(201).send();
});

app.get('/statement', verifyAccount, (request, response) => {
  const { customer } = request;
  return response.json(customer.statement);
});

app.get('/statement/date', verifyAccount, (request, response) => {
  const { customer } = request;
  const { date } = request.query;

  const dateFormat = new Date(`${date} 00:00`);

  const statement = customer.statement.filter(
    state =>
      state.created_at.toDateString() === new Date(dateFormat).toDateString(),
  );

  return response.json(statement);
});

app.get('/balance', verifyAccount, (request, response) => {
  const { customer } = request;

  const balance = getBalance(customer.statement);

  return response.json(balance);
});

app.listen(PORT, () => {
  console.log(
    `👀 Server http://localhost:${PORT}/swagger is being watched - 'Quis custodiet ipsos custodes? 🤔'!`,
  );
});